core: Modifications were made to be able to add important files with namespace, and a refactoring of the portal's homepage.
